# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Sanket Matre Portfolio
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterBody {
  fullName: string;
  email: string;
  /** @minLength 6 */
  password: string;
  confirmPassword: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  fullName: string;
  email: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface ContactBody {
  name: string;
  email: string;
  message: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}