# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Sanket Matre Portfolio
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const registerBodyPasswordMin = 6;

export const RegisterBody = zod.object({
  fullName: zod.string(),
  email: zod.string().email(),
  password: zod.string().min(registerBodyPasswordMin),
  confirmPassword: zod.string(),
});

/**
 * @summary Login with email and password
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    fullName: zod.string(),
    email: zod.string(),
  }),
});

/**
 * @summary Get the current authenticated user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  fullName: zod.string(),
  email: zod.string(),
});

/**
 * @summary Logout the current user
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Submit a contact form message
 */
export const SubmitContactBody = zod.object({
  name: zod.string(),
  email: zod.string().email(),
  message: zod.string(),
});