# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Sanket Matre Portfolio
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from "./userProfile";

export interface AuthResponse {
  token: string;
  user: UserProfile;
}