# lib/api-zod/src/generated/types/messageResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Sanket Matre Portfolio
 * OpenAPI spec version: 0.1.0
 */

export interface MessageResponse {
  message: string;
}