# lib/api-zod/src/generated/types/registerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Sanket Matre Portfolio
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterBody {
  fullName: string;
  email: string;
  /** @minLength 6 */
  password: string;
  confirmPassword: string;
}